+++ test/api/initiation.ts
@@ -7,12 +7,12 @@ describe('Initiat', () => {
     new Confetti({ apiKey: 'my-key' })
   })
 
-  test('should fail to call due on Confetti api instance due to missing api key', async () => {
+  test('should fail to call on a Confetti api instance with no credentials', async () => {
     const confetti = new Confetti()
     await assert.rejects(
       () => confetti.events.findAll(),
       (error: unknown) => {
-        return error instanceof Error && error.message.includes('missing_api_key')
+        return error instanceof Error && error.message.includes('missing_credentials')
       },
     )
   })